Fixed lint issues

+++ src/uproot/_dask.py
@@ -632,9 +632,7 @@ def _get_dask_array(
         assert steps_per_file is not unset  # either assigned or assumed to be 1
         total_files = len(ttrees)
         total_entries = sum(ttree.num_entries for ttree in ttrees)
-        step_size = max(
-            1, int(math.ceil(total_entries / (total_files * steps_per_file)))
-        )
+        step_size = max(1, math.ceil(total_entries / (total_files * steps_per_file)))
 
     if count == 0:
         raise ValueError(
@@ -672,7 +670,7 @@ def _get_dask_array(
         )
         step_sum += int(ttree_step)
 
-    entry_step = int(round(step_sum / len(ttrees)))
+    entry_step = round(step_sum / len(ttrees))
     assert entry_step >= 1
 
     for key in common_keys:
@@ -1490,9 +1488,7 @@ def _get_dak_array(
         assert steps_per_file is not unset  # either assigned or assumed to be 1
         total_files = len(ttrees)
         total_entries = sum(ttree.num_entries for ttree in ttrees)
-        step_size = max(
-            1, int(math.ceil(total_entries / (total_files * steps_per_file)))
-        )
+        step_size = max(1, math.ceil(total_entries / (total_files * steps_per_file)))
 
     if count == 0:
         raise ValueError(
@@ -1528,7 +1524,7 @@ def _get_dak_array(
         )
         step_sum += int(ttree_step)
 
-    entry_step = int(round(step_sum / len(ttrees)))
+    entry_step = round(step_sum / len(ttrees))
 
     divisions = [0]
     partition_args = []
