+++ src/uproot/models/RNTuple.py
@@ -166,7 +166,7 @@ def _num_entries_for(in_ntuple, target_num_bytes, filter_name):
     if total_bytes == 0:
         num_entries = 0
     else:
-        num_entries = int(round(target_num_bytes * total_entries / total_bytes))
+        num_entries = round(target_num_bytes * total_entries / total_bytes)
     if num_entries <= 0:
         return 1
     else:
